debugged create_input.ipynb (#1)
There is one optional fix here. I noticed that samples are renamed according to the last element in the string. This is fine if samples are indexed such that they end a unique integer/letter, but this created problems for a dataset I was working with. In this PR, samples are instead labeled according to their full, original name.